github

+++ controller/lib/adminController.js
@@ -3,38 +3,51 @@ const util = require('../../util')
 const objectId = require('mongodb').ObjectId
 
 module.exports = {
-    async getAdmin(ctx, next){
+    async getAdmin(ctx, next) {
         let params = ctx.query || ctx.request.query
         if (!params.user_id) {
             return ctx.body = util.json(0, {
                 msg: '参数user_id不能为空'
             })
         } else {
-            let data = await DB.find('users','user_id',[params.user_id]);
+            let data = await DB.find('users', 'user_id', [params.user_id]);
             ctx.body = util.json(1, {
-                is_admin:data[0].is_admin,
-                super_admin:data[0].super_admin || 0
+                is_admin: data[0].is_admin,
+                super_admin: data[0].super_admin || 0
             })
         }
     },
-    async delAccount(ctx,next){
+    async delAccount(ctx, next) {
         let params = ctx.body || ctx.request.body
-        if(!params.accountId){
-            ctx.body = util.json(0,{
-                msg:'删除的账户ID不能为空'
+        if (!params.accountId) {
+            ctx.body = util.json(0, {
+                msg: '删除的账户ID不能为空'
             })
             return
         }
-        await DB.remove('users','user_id',[params.accountId])
-        await DB.remove('article','user_id',[params.accountId])
-        ctx.body = util.json(1)
+        try {
+            let data = await DB.find('users', 'user_id', [params.accountId]);
+            if (data.length) {
+                await util.removeImg(ctx,'accountId')
+                await DB.remove('users', 'user_id', [params.accountId])
+                let articleList = DB.find('article','user_id',[params.accountId])
+                if(articleList.length){
+                    await DB.remove('article', 'user_id', [params.accountId])
+                }
+                ctx.body = util.json(1)
+            }
+        } catch (err) {
+            ctx.body = util.json(0,{
+                msg:err
+            })
+        }
     },
-    async adminOprate(ctx,next){
+    async adminOprate(ctx, next) {
         let params = ctx.body || ctx.request.body
-        let data = await DB.find('users','user_id',[params.id])
-        let isAdmin = data[0].is_admin > 0? 0 : 1
-        let res = await DB.update('users',`is_admin="${isAdmin}"`,'user_id',[params.id])
-        if(res){
+        let data = await DB.find('users', 'user_id', [params.id])
+        let isAdmin = data[0].is_admin > 0 ? 0 : 1
+        let res = await DB.update('users', `is_admin="${isAdmin}"`, 'user_id', [params.id])
+        if (res) {
             ctx.body = util.json(1)
         }
     }
